Marshal enumerable to list
Supports both `Run(string, object)` and `Run(string, IDictionary<string, object>)`
When converting from object to `IDictionary<string, object>`, if the object implements `IEnumerable`, then it would call `GetEnumerator` rather than try to convert the class as a `IDictionary<string, object>`.
